+++ spec/dummy_app/app/models/concerns/prefix_versions_inspect_with_count.rb
@@ -1,0 +1,8 @@
+# frozen_string_literal: true
+
+module PrefixVersionsInspectWithCount
+  def inspect
+    "#{length} versions:\n" +
+      super
+  end
+end