+++ spec/dummy_app/app/models/fruit.rb
@@ -2,6 +2,6 @@
 
 class Fruit < ActiveRecord::Base
   if ENV["DB"] == "postgres" || JsonVersion.table_exists?
-    has_paper_trail class_name: "JsonVersion"
+    has_paper_trail versions: { class_name: "JsonVersion" }
   end
 end